Move validator to search directory
