+++ src/lsdb/catalog/catalog.py
@@ -13,11 +13,11 @@ from hipscat.pixel_math.polygon_filter import SphericalCoordinates
 from lsdb import io
 from lsdb.catalog.association_catalog import AssociationCatalog
 from lsdb.catalog.dataset.healpix_dataset import HealpixDataset
-from lsdb.catalog.validators import CoordinatesValidator
 from lsdb.core.crossmatch.abstract_crossmatch_algorithm import AbstractCrossmatchAlgorithm
 from lsdb.core.crossmatch.crossmatch_algorithms import BuiltInCrossmatchAlgorithm
 from lsdb.core.search.cone_search import cone_filter
 from lsdb.core.search.polygon_search import get_cartesian_polygon, polygon_filter
+from lsdb.core.search.validators import CoordinatesValidator
 from lsdb.dask.crossmatch_catalog_data import crossmatch_catalog_data
 from lsdb.dask.divisions import get_pixels_divisions
 from lsdb.dask.join_catalog_data import join_catalog_data_on, join_catalog_data_through
