+++ src/lsdb/core/search/validators.py
